creating schema

+++ cv-api/db/migrate/20220320213832_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[6.1]
   def change
     create_table :products do |t|
       t.string :name_product
-      t.double :price
+      t.integer :price
 
       t.timestamps
     end
